# web/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateUserInput": {
        "properties": {
          "display_name": {
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "minLength": 12,
            "type": "string"
          },
          "role": {
            "description": "The role granted in tenant",
            "type": "string"
          },
          "superadmin": {
            "default": false,
            "type": "boolean"
          },
          "tenant": {
            "description": "Optional tenant granted role in the create transaction",
            "type": "string"
          },
          "workspace_id": {
            "description": "Optional workspace the membership is confined to",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "CreatedUser": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "is_superadmin": {
            "type": "boolean"
          },
          "membership": {
            "description": "The first membership, granted in the create transaction when tenant was given",
            "properties": {
              "id": {
                "format": "uuid",
                "type": "string"
              },
              "role": {
                "type": "string"
              },
              "tenant": {
                "type": "string"
              },
              "user_id": {
                "format": "uuid",
                "type": "string"
              },
              "workspace_id": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "status": {
            "enum": [
              "active",
              "inactive"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "status",
          "is_superadmin",
          "created_at"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "User": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "is_superadmin": {
            "type": "boolean"
          },
          "status": {
            "enum": [
              "active",
              "inactive"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "status",
          "is_superadmin",
          "created_at"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Opaque session (olvs_) or API (olvk_) token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST contract for the Olivares AI control plane. Authentication is an opaque bearer token (session olvs_… or API key olvk_…). Tenant is resolved from a bound token or the X-Olivares-Tenant header.",
    "license": {
      "name": "AGPL-3.0-only"
    },
    "title": "Olivares AI control plane API",
    "version": "v1",
    "x-stability-policy": "https://olivares.ai/docs"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/users": {
      "get": {
        "description": "Lists the user accounts of the deployment. Superadmin only.",
        "operationId": "listUsers",
        "parameters": [
          {
            "description": "Maximum number of items to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Pagination cursor from a previous response.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "cursor": {
                      "type": "string"
                    },
                    "has_more": {
                      "type": "boolean"
                    },
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/User"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "items",
                    "has_more"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict / setup required"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List users (superadmin)",
        "tags": [
          "users"
        ],
        "x-required-permission": "user:read",
        "x-stability": "stable"
      },
      "post": {
        "description": "Creates one user account. Superadmin only.",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedUser"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict / setup required"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a user (superadmin)",
        "tags": [
          "users"
        ],
        "x-required-permission": "user:write",
        "x-stability": "stable"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "this engine",
      "url": "/"
    }
  ],
...
}